# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cyland API",
    "version": "1.0.0",
    "description": "Next.js App Router 기반 API 문서"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/post-story": {
      "post": {
        "tags": [
          "Story"
        ],
        "summary": "스토리 생성",
        "description": "관리자 인증 필요.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "content"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "스토리 이름"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "생성된 스토리 (id, name, content, active)"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "인증 필요"
          },
          "500": {
            "description": "서버 에러"
          }
        }
      }
    },
...
    "/api/put-story": {
      "put": {
        "tags": [
          "Story"
        ],
        "summary": "스토리 수정",
        "description": "관리자 인증 필요.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "name",
                  "content"
                ],
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string",
                    "description": "스토리 이름"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 id, name, content"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "인증 필요"
          },
          "404": {
            "description": "Story not found"
          },
          "500": {
            "description": "서버 에러"
          }
        }
      }
    },
...
  },
...
}